requirements.txt
bcrypt==3.2.2
blinker==1.6.2
capstone==5.0.0rc2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.7
colorama==0.4.6
colored-traceback==0.3.0
cryptography==36.0.2
exceptiongroup==1.2.0
Flask==2.3.3
idna==3.3
iniconfig==2.0.0
intervaltree==3.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
paramiko==2.11.0
pathlib2==2.3.7.post1
pluggy==1.3.0
plumbum==1.7.2
psutil==5.9.1
psycopg2-binary==2.9.9
pwntools==4.8.0
pycparser==2.21
pycryptodome==3.15.0
pyelftools==0.28
Pygments==2.12.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
pytest==7.4.3
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
# Exclude pywin32 for non-Windows platforms
pywin32==304 ; sys_platform == 'win32'
requests==2.28.1
ROPGadget==6.8
rpyc==5.1.0
shortuuid==1.0.11
six==1.16.0
sortedcontainers==2.4.0
tomli==2.0.1
unicorn==2.0.0
urllib3==1.26.10
Werkzeug==2.3.7
